+++ core/src/com/gsma/rcs/service/api/IPCallImpl.java
@@ -180,7 +180,7 @@ public class IPCallImpl extends IIPCall.Stub implements IPCallStreamingSessionLi
     public int getState() {
         IPCallSession session = mIPCallService.getIPCallSession(mCallId);
         if (session == null) {
-            return mPersistentStorage.getState();
+            return mPersistentStorage.getState().toInt();
         }
         SipDialogPath dialogPath = session.getDialogPath();
         if (dialogPath != null && dialogPath.isSessionEstablished()) {
@@ -202,7 +202,7 @@ public class IPCallImpl extends IIPCall.Stub implements IPCallStreamingSessionLi
     public int getReasonCode() {
         IPCallSession session = mIPCallService.getIPCallSession(mCallId);
         if (session == null) {
-            return mPersistentStorage.getReasonCode();
+            return mPersistentStorage.getReasonCode().toInt();
         }
         return ReasonCode.UNSPECIFIED.toInt();
     }
